refactor: modularize endRoundCommand into focused handlers
Split the 428-line endRoundCommand.ts into an endRound/ directory with
focused handlers:
- types.ts: Type definitions for handler results
- gameEnd.ts: Game end detection and scoring
- timeTransition.ts: Day/night toggle and dawn ruins reveal
- manaReset.ts: Mana source reset
- offerRefresh.ts: Unit, AA, spell, and monastery offer refresh
- playerRoundReset.ts: Deck shuffle, unit ready, turn state reset
- tacticsSetup.ts: Tactics collection and selection phase setup
- index.ts: Main orchestrator

Each handler has a single responsibility and can be tested in isolation.
